backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0
reportlab==4.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
